# server/openAPI spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Medium Clone API",
    "version": "1.0.0",
    "description": "A node.js Express API"
  },
  "paths": {
...
    "/api/v1/like": {
      "post": {
        "summary": "Like a Post",
        "tags": [
          "Like"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "userId",
                  "blogId"
                ],
                "properties": {
                  "userId": {
                    "type": "integer",
                    "description": "ID of the user liking the blog"
                  },
                  "blogId": {
                    "type": "integer",
                    "description": "ID of the blog post to like"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Entry Created Successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "userId": {
                          "type": "integer"
                        },
                        "blogId": {
                          "type": "integer"
                        }
                      }
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error Liking the Post",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/comment": {
      "post": {
        "summary": "Add comment to a post",
        "tags": [
          "Comment"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "userId",
                  "blogId",
                  "body"
                ],
                "properties": {
                  "userId": {
                    "type": "integer",
                    "description": "ID of the user commenting on the blog"
                  },
                  "blogId": {
                    "type": "integer",
                    "description": "ID of the blog post to comment on"
                  },
                  "body": {
                    "type": "string",
                    "description": "Content of the comment"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Entry Created Successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "userId": {
                          "type": "integer"
                        },
                        "blogId": {
                          "type": "integer"
                        },
                        "body": {
                          "type": "string"
                        }
                      }
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error commenting on the post",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/saveblog": {
      "post": {
        "summary": "Save a blog",
        "tags": [
          "Savedblog"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "userId",
                  "blogId"
                ],
                "properties": {
                  "userId": {
                    "type": "integer",
                    "description": "ID of the user saving the blog post"
                  },
                  "blogId": {
                    "type": "integer",
                    "description": "ID of the blog post to be saved"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully saved the post",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "userId": {
                          "type": "integer"
                        },
                        "blogId": {
                          "type": "integer"
                        }
                      }
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error saving the blog",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/deletelike": {
      "delete": {
        "summary": "Remove the like",
        "tags": [
          "Like"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "likeId"
                ],
                "properties": {
                  "likeId": {
                    "type": "integer",
                    "description": "ID of the post like to be removed"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully deleted the like",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer"
                        }
                      }
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error deleting the like",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/deletecomment": {
      "delete": {
        "summary": "Remove the comment",
        "tags": [
          "Comment"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "commentId"
                ],
                "properties": {
                  "commentId": {
                    "type": "integer",
                    "description": "ID of the comment to be removed"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully deleted the comment",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer"
                        }
                      }
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error deleting the comment",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/deletesave": {
      "delete": {
        "summary": "Remove the saved post",
        "tags": [
          "Savedblog"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "savedblogId"
                ],
                "properties": {
                  "savedblogId": {
                    "type": "integer",
                    "description": "ID of the saved blog to be removed"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully deleted the saved blog",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer"
                        }
                      }
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error deleting the saved blog",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}